floats en todo y correcciones a las entidades

+++ proyecto_final/src/main/java/joystickmx/itson/persistencia/ItemCarrito.java
@@ -87,11 +87,11 @@ public class ItemCarrito implements Serializable {
     }
     
     @Transient
-    public Double getSubtotal() {
+    public Float getSubtotal() {
         if (videojuego != null && cantidad != null) {
             return videojuego.getPrecio() * cantidad;
         }
-        return 0.0;
+        return 0.0f;
     }
     
     @Override
